+++ xcube/core/gen2/response.py
@@ -44,7 +44,7 @@ class GenericCubeGeneratorResult(Generic[R], JsonObject):
                  message: Optional[str] = None,
                  output: Optional[Sequence[str]] = None,
                  traceback: Optional[Sequence[str]] = None,
-                 dependencies: Optional[Dict[str, str]] = None):
+                 versions: Optional[Dict[str, str]] = None):
         assert_instance(status, str, name='status')
         assert_in(status, STATUS_IDS, name='status')
         self.status = status
@@ -53,7 +53,7 @@ class GenericCubeGeneratorResult(Generic[R], JsonObject):
         self.message = message if message else None
         self.output = list(output) if output else None
         self.traceback = list(traceback) if traceback else None
-        self.dependencies = dependencies if dependencies else None
+        self.versions = versions if versions else None
 
     def derive(self,
                /,
@@ -63,14 +63,14 @@ class GenericCubeGeneratorResult(Generic[R], JsonObject):
                message: Optional[str] = None,
                output: Optional[Sequence[str]] = None,
                traceback: Optional[Sequence[str]] = None,
-               dependencies: Optional[Dict[str, str]] = None) -> R:
+               versions: Optional[Dict[str, str]] = None) -> R:
         return self.__class__(status or self.status,
                               status_code=status_code or self.status_code,
                               result=result or self.result,
                               message=message or self.message,
                               output=output or self.output,
                               traceback=traceback or self.traceback,
-                              dependencies=dependencies or self.dependencies)
+                              versions=versions or self.versions)
 
     @classmethod
     @abstractmethod
@@ -91,7 +91,7 @@ class GenericCubeGeneratorResult(Generic[R], JsonObject):
                 message=JsonStringSchema(),
                 output=JsonArraySchema(items=JsonStringSchema()),
                 traceback=JsonArraySchema(items=JsonStringSchema()),
-                dependencies=JsonObjectSchema(additional_properties=True)
+                versions=JsonObjectSchema(additional_properties=True)
             ),
             required=['status'],
             additional_properties=True,